+++ modules/core/src/test/java/test/jts/perf/operation/overlayng/RandomPolygonOverlayFuzzer.java
@@ -123,7 +123,7 @@ public class RandomPolygonOverlayFuzzer {
  
   private String stats() {
     return String.format("\nTest %d: %s   (# errs: %d = %d%%)\n", testIndex, testDesc, 
-        errCount, (int) (100 * errCount) / testIndex );
+        errCount, 100 * errCount / testIndex );
   }
 
   private void overlayNG(Geometry poly1, Geometry poly2) {
